feat(data-designer): drop legacy CLI verbs
