+++ plugins/nemo-data-designer/src/nemo_data_designer_plugin/cli/inputs.py
@@ -74,18 +74,12 @@ def _spec_from_builder(config_source: str, num_records: int) -> Iterator[str]:
 
 
 def apply_preview_cli_overrides(group: typer.Typer) -> None:
-    """Replace ``preview run`` / ``preview submit`` with friendly wrappers."""
-    _replace_function_run(group)
+    """Replace ``nemo data-designer preview`` with friendly wrapper that matches upstream library."""
     _replace_function_submit(group)
 
 
 def apply_create_cli_overrides(group: typer.Typer) -> None:
-    """Replace ``create run`` / ``create submit`` with friendly wrappers.
-
-    Jobs also have an ``explain`` verb that prints schemas; it's not affected
-    here because it doesn't take a config-source input.
-    """
-    _replace_job_run(group)
+    """Replace ``nemo data-designer create`` with friendly wrapper that matches upstream library."""
     _replace_job_submit(group)
 
 
@@ -101,40 +95,11 @@ def _pluck_callback(group: typer.Typer, verb: str) -> Callable[..., None]:
     return callback
 
 
-def _replace_function_run(group: typer.Typer) -> None:
-    original = _pluck_callback(group, "run")
-
-    @group.command("run")
-    def run(
-        typer_ctx: typer.Context,
-        config_source: str = typer.Argument(..., metavar="[CONFIG_SOURCE]", help=_CONFIG_SOURCE_HELP),
-        num_records: int = typer.Option(
-            DEFAULT_NUM_RECORDS, "--num-records", "-n", help="Number of records to generate.", min=1
-        ),
-        workspace: str = typer.Option(
-            "default", "--workspace", "-w", help="Workspace identity passed to the function as ctx.workspace."
-        ),
-        non_interactive: bool = typer.Option(False, "--non-interactive", help=_NON_INTERACTIVE_HELP),
-        save_results: bool = typer.Option(False, "--save-results", help=_SAVE_RESULTS_HELP),
-        artifact_path: str | None = typer.Option(None, "--artifact-path", "-o", help=_ARTIFACT_PATH_HELP),
-    ) -> None:
-        with _spec_from_builder(config_source, num_records) as spec:
-            original(
-                typer_ctx,
-                spec=spec,
-                spec_file=None,
-                workspace=workspace,
-                non_interactive=non_interactive,
-                save_results=save_results,
-                artifact_path=artifact_path,
-            )
-
-
 def _replace_function_submit(group: typer.Typer) -> None:
-    original = _pluck_callback(group, "submit")
+    original = _pluck_callback(group, "preview")
 
-    @group.command("submit")
-    def submit(
+    @group.command("preview")
+    def preview(
         typer_ctx: typer.Context,
         config_source: str = typer.Argument(..., metavar="[CONFIG_SOURCE]", help=_CONFIG_SOURCE_HELP),
         num_records: int = typer.Option(DEFAULT_NUM_RECORDS, "--num-records", "-n", min=1),
@@ -161,26 +126,11 @@ def _replace_function_submit(group: typer.Typer) -> None:
             )
 
 
-def _replace_job_run(group: typer.Typer) -> None:
-    original = _pluck_callback(group, "run")
-
-    @group.command("run")
-    def run(
-        typer_ctx: typer.Context,
-        config_source: str = typer.Argument(..., metavar="[CONFIG_SOURCE]", help=_CONFIG_SOURCE_HELP),
-        num_records: int = typer.Option(
-            DEFAULT_NUM_RECORDS, "--num-records", "-n", help="Number of records to generate.", min=1
-        ),
-    ) -> None:
-        with _spec_from_builder(config_source, num_records) as spec:
-            original(typer_ctx, spec=spec, spec_file=None, config=None, config_file=None)
-
-
 def _replace_job_submit(group: typer.Typer) -> None:
-    original = _pluck_callback(group, "submit")
+    original = _pluck_callback(group, "create")
 
-    @group.command("submit")
-    def submit(
+    @group.command("create")
+    def create(
         typer_ctx: typer.Context,
         config_source: str = typer.Argument(..., metavar="[CONFIG_SOURCE]", help=_CONFIG_SOURCE_HELP),
         num_records: int = typer.Option(DEFAULT_NUM_RECORDS, "--num-records", "-n", min=1),
